starboard

+++ moderation/src/main/java/pl/fratik/moderation/commands/ReasonCommand.java
@@ -17,8 +17,15 @@
 
 package pl.fratik.moderation.commands;
 
+import net.dv8tion.jda.api.Permission;
+import net.dv8tion.jda.api.entities.User;
+import net.dv8tion.jda.api.interactions.commands.DefaultMemberPermissions;
 import org.jetbrains.annotations.NotNull;
-import pl.fratik.core.entity.*;
+import pl.fratik.core.command.NewCommandContext;
+import pl.fratik.core.entity.Akcja;
+import pl.fratik.core.entity.Kara;
+import pl.fratik.core.entity.Schedule;
+import pl.fratik.core.entity.ScheduleDao;
 import pl.fratik.core.util.DurationUtil;
 import pl.fratik.moderation.entity.AutoAkcja;
 import pl.fratik.moderation.entity.Case;
@@ -26,10 +33,6 @@ import pl.fratik.moderation.entity.CaseDao;
 import pl.fratik.moderation.utils.ReasonUtils;
 
 import java.time.Instant;
-import java.util.Arrays;
-import java.util.LinkedHashMap;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 public class ReasonCommand extends ModerationCommand {
     private final ScheduleDao scheduleDao;
@@ -39,60 +42,62 @@ public class ReasonCommand extends ModerationCommand {
         super(true);
         this.scheduleDao = scheduleDao;
         this.caseDao = caseDao;
-        name = "reason";
-        category = CommandCategory.MODERATION;
-        uzycieDelim = " ";
-        LinkedHashMap<String, String> hmap = new LinkedHashMap<>();
-        hmap.put("caseid", "integer");
-        hmap.put("powod", "string");
-        hmap.put("[...]", "string");
-        uzycie = new Uzycie(hmap, new boolean[] {true, true, false});
-        aliases = new String[] {"powod", "powód"};
+        name = "powod";
+        permissions = DefaultMemberPermissions.enabledFor(Permission.MODERATE_MEMBERS);
+        usage = "<numer_sprawy:int> <powod:string>";
     }
 
     @Override
-    public boolean execute(@NotNull CommandContext context) {
-        Integer caseId = (Integer) context.getArgs()[0];
-        String reason = Arrays.stream(Arrays.copyOfRange(context.getArgs(), 1, context.getArgs().length))
-                .map(e -> e == null ? "" : e).map(Objects::toString).collect(Collectors.joining(uzycieDelim));
+    public void execute(@NotNull NewCommandContext context) {
+        int caseId = context.getArguments().get("numer_sprawy").getAsInt();
+        String reason = context.getArguments().get("powod").getAsString();
+        context.defer(false);
         Case aCase = caseDao.getLocked(CaseDao.getId(context.getGuild(), caseId));
         if (aCase == null) {
-            context.reply(context.getTranslated("reason.invalid.case"));
-            return false;
+            context.sendMessage(context.getTranslated("reason.invalid.case"));
+            return;
         }
         try {
             DurationUtil.Response durationResp;
             try {
                 durationResp = DurationUtil.parseDuration(reason);
             } catch (IllegalArgumentException e) {
-                context.reply(context.getTranslated("reason.max.duration"));
-                return false;
+                context.sendMessage(context.getTranslated("reason.max.duration"));
+                return;
             }
             String powod = durationResp.getTekst();
             Instant akcjaDo = durationResp.getDoKiedy();
             if (powod.equals("")) {
-                context.reply(context.getTranslated("reason.reason.empty"));
-                return false;
+                context.sendMessage(context.getTranslated("reason.reason.empty"));
+                return;
             }
             aCase.setIssuerId(context.getSender().getIdLong());
             aCase.getFlagi().remove(Case.Flaga.NOBODY); // usuwa -n, -n może zostać ponownie dodane w parseFlags
             if (akcjaDo != null) {
-                if (aCase.getValidTo() != null) {
-                    for (Schedule schedule : scheduleDao.getByDate(Instant.from(aCase.getValidTo()).toEpochMilli())) {
-                        if (schedule.getContent() instanceof AutoAkcja &&
-                                ((AutoAkcja) schedule.getContent()).getCaseId() == aCase.getCaseNumber())
-                            scheduleDao.delete(Integer.toString(schedule.getId()));
+                if (aCase.getType() == Kara.MUTE) {
+                    try {
+                        context.getGuild().timeoutUntil(User.fromId(aCase.getUserId()), akcjaDo).complete();
+                    } catch (Exception e) {
+                        context.sendMessage(context.getTranslated("reason.timeout.change.error"));
+                        return;
                     }
+                } else {
+                    if (aCase.getValidTo() != null) {
+                        for (Schedule schedule : scheduleDao.getByDate(Instant.from(aCase.getValidTo()).toEpochMilli())) {
+                            if (schedule.getContent() instanceof AutoAkcja &&
+                                    ((AutoAkcja) schedule.getContent()).getCaseId() == aCase.getCaseNumber())
+                                scheduleDao.delete(Integer.toString(schedule.getId()));
+                        }
+                    }
+                    //noinspection ConstantConditions - ustawiane wyżej
+                    scheduleDao.save(scheduleDao.createNew(akcjaDo.toEpochMilli(), Long.toUnsignedString(aCase.getIssuerId()), Akcja.EVENT,
+                            new AutoAkcja(aCase.getCaseNumber(), aCase.getType().opposite(), Long.toUnsignedString(aCase.getGuildId()))));
                 }
                 aCase.setValidTo(akcjaDo);
-                //noinspection ConstantConditions - ustawiane wyżej
-                scheduleDao.save(scheduleDao.createNew(akcjaDo.toEpochMilli(), Long.toUnsignedString(aCase.getIssuerId()), Akcja.EVENT,
-                        new AutoAkcja(aCase.getCaseNumber(), aCase.getType().opposite(), Long.toUnsignedString(aCase.getGuildId()))));
             }
             ReasonUtils.parseFlags(aCase, powod, Case.Flaga.SILENT);
             caseDao.save(aCase);
-            context.send(context.getTranslated("reason.success"));
-            return true;
+            context.sendMessage(context.getTranslated("reason.success"));
         } finally {
             caseDao.unlock(aCase);
         }
